Stop registering MCP tools in the browser
They execute in the runtime now, so a second registration in the surface
would offer the model every tool twice. The result rendering survives:
`forDisplay` moves to a shared module and the transcript's fallback
draws it, which is no longer the exception but the ordinary path for
every MCP call.
