+++ app/src/lib/plugins/tool-result.ts
@@ -1,0 +1,45 @@
+/**
+ * A tool result, as something worth looking at.
+ *
+ * MCP says a text part, and vendors fill it with anything from plain markdown to a JSON envelope
+ * with the markdown inside one field. Markdown is drawn as markdown, a JSON wrapper is unwrapped to
+ * the markdown it was hiding, and anything else is fenced as JSON. Nothing is discarded.
+ */
+export function forDisplay(text: string): string {
+  const trimmed = text.trim();
+  if (!trimmed.startsWith("{") && !trimmed.startsWith("[")) return text;
+
+  let parsed: unknown;
+  try {
+    parsed = JSON.parse(trimmed);
+  } catch {
+    // Not JSON after all. Draw what the server sent.
+    return text;
+  }
+
+  if (parsed && typeof parsed === "object" && !Array.isArray(parsed)) {
+    const entries = Object.entries(parsed as Record<string, unknown>);
+    // The field carrying the answer, told apart from the ones carrying bookkeeping: Slack sends
+    // `{ results, pagination_info }`. The rest is kept below rather than dropped.
+    const markdown = entries
+      .filter(
+        ([, value]) =>
+          typeof value === "string" &&
+          (value.includes("\n#") || value.startsWith("#")),
+      )
+      .sort((a, b) => String(b[1]).length - String(a[1]).length)[0];
+
+    if (markdown) {
+      const rest = entries.filter(([key]) => key !== markdown[0]);
+      const body = String(markdown[1]);
+      if (rest.length === 0) return body;
+      return `${body}\n\n\`\`\`json\n${JSON.stringify(
+        Object.fromEntries(rest),
+        null,
+        2,
+      )}\n\`\`\``;
+    }
+  }
+
+  return `\`\`\`json\n${JSON.stringify(parsed, null, 2)}\n\`\`\``;
+}
